feat: split the raffle page into prizes, draw, and winners panels
